Added android specific implementation of LocatioProvider

+++ app/src/main/java/talosdev/clean/features/location/model/Location.java
@@ -5,10 +5,10 @@ import com.google.auto.value.AutoValue;
 @AutoValue
 public abstract class Location {
 
-    public abstract long latitude();
-    public abstract long longitude();
+    public abstract double latitude();
+    public abstract double longitude();
 
-    public static Location create(long latitude, long longitude) {
+    public static Location create(double latitude, double longitude) {
         return new AutoValue_Location(latitude, longitude);
     }
